+++ libs/spark/src/Unstable_FormControl/Unstable_FormControl.tsx
@@ -2,6 +2,8 @@ import React, { ElementType, forwardRef } from 'react';
 import MuiFormControl, {
   FormControlProps as MuiFormControlProps,
 } from '@material-ui/core/FormControl';
+import clsx from 'clsx';
+import makeStyles from '../makeStyles';
 import { OverridableComponent, OverrideProps } from '../utils';
 
 export interface Unstable_FormControlTypeMap<
@@ -26,11 +28,29 @@ export type Unstable_FormControlProps<
 
 export type Unstable_FormControlClassKey = 'root';
 
+const useStyles = makeStyles<Unstable_FormControlClassKey>(
+  {
+    /** Styles applied to the root element. */
+    root: {
+      gap: 8,
+    },
+  },
+  { name: 'MuiSparkUnstable_FormControl' }
+);
+
 const Unstable_FormControl: OverridableComponent<Unstable_FormControlTypeMap> = forwardRef(
   function Unstable_FormControl(props, ref) {
-    const { color: _color, ...other } = props;
+    const { classes: classesProp, color: _color, ...other } = props;
 
-    return <MuiFormControl ref={ref} {...other} />;
+    const classes = useStyles();
+
+    return (
+      <MuiFormControl
+        classes={{ root: clsx(classes.root, classesProp?.root) }}
+        ref={ref}
+        {...other}
+      />
+    );
   }
 );